20220627----yixiu---------游戏api登录

+++ indo-game/game-api/src/main/java/com/indo/game/pojo/dto/cq/CqEndroundDataCallBackReq.java
@@ -12,7 +12,7 @@ public class CqEndroundDataCallBackReq extends CqCallBackParentReq {
     @JSONField(name = "mtcode")
     private String mtcode;//混合碼
     @JSONField(name = "amount")
-    private BigDecimal amount;//金額
+    private Double amount;//金額
     @JSONField(name = "eventtime")
     private String eventtime;//事件時間 time.RFC3339
     @JSONField(name = "validbet")
